Extract compression-related tests to separate compressed.rs file
- Created new compressed.rs test file for all compression-related tests
- Moved test_format_detection from read_tests.rs to compressed.rs
- Moved test_write_file_compressed from write_tests.rs to compressed.rs
- Added new tests:
  - test_auto_format_detection: Tests automatic format detection by file extension
  - test_load_zipped_reader: Tests in-memory compressed data loading
- Fixed Instance comparison to use abs_diff_eq for order-independent comparison
- Removed unnecessary imports from read_tests.rs and write_tests.rs
- All 250 tests pass successfully
