+++ rust/ommx/src/mps/tests/read_tests.rs
@@ -1,8 +1,6 @@
 use super::super::*;
 use super::MPS_COMPLEX;
 use std::collections::BTreeMap;
-use std::io::Write;
-use tempdir::TempDir;
 
 // Test basic MPS parsing
 #[test]
@@ -194,43 +192,6 @@ ENDATA
     assert_eq!(instance.sense(), crate::v1::instance::Sense::Maximize);
 }
 
-// Test format detection (compressed vs uncompressed)
-#[test]
-fn test_format_detection() {
-    const MPS_CONTENT: &str = r#"NAME TestProblem
-ROWS
- N  OBJ
- L  R1
-COLUMNS
-    X1        OBJ                 1
-    X1        R1                  1
-RHS
-    RHS1      R1                  5
-BOUNDS
- UP BND1      X1                  4
-ENDATA
-"#;
-
-    let temp_dir = TempDir::new("test_mps_format_detection").unwrap();
-    let temp_dir_path = temp_dir.path();
-    let uncompressed_path = temp_dir_path.join("test.mps");
-    let compressed_path = temp_dir_path.join("test.mps.gz");
-
-    // Create uncompressed file
-    std::fs::write(&uncompressed_path, MPS_CONTENT).unwrap();
-
-    // Create compressed file
-    {
-        let file = std::fs::File::create(&compressed_path).unwrap();
-        let mut encoder = flate2::write::GzEncoder::new(file, flate2::Compression::default());
-        encoder.write_all(MPS_CONTENT.as_bytes()).unwrap();
-        encoder.finish().unwrap();
-    }
-
-    let uncompressed = load_file(&uncompressed_path).unwrap();
-    let compressed = load_file(&compressed_path).unwrap();
-    assert_eq!(compressed, uncompressed);
-}
 
 // Test complex MPS with all constraint types
 #[test]
